+++ CHANGELOG.md
@@ -13,6 +13,8 @@ That's all, matcher will stop when detects next line started with `###` match
 * The map will now remove dropped items & orbs after game finish
 * Fixed some bugs that trigger events on physical action (for example spectator gui opens on redstone block)
 * Fixed fast respawn bug
+* Now players can´t interact with armorstands when they are in the arena
+* Now players can´t destroy itemframes, paintings, and armorstands when they are in the arena
 
 ### 4.2.2 Release (14.06.2019 - 13.07.2019)
 * Fixed setup menu completed/not completed values were buggy